requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
azure-identity
pyodbc
azure-core==1.24.2
azure-storage-blob==12.9.0
azure-storage-queue==12.5.0